refactor(product): move internal DTOs from order to product module
- Move ProcessOrderItemsRequest to product.DTO.internal
- Move OrderItemRequest to product.DTO.internal
- Update all references in ProductController, ProductClient, ProductService, OrderService
- Update ProductServiceTest and OrderServiceTest
- Remove 'Made with Bob' comments

This resolves cross-module DTO coupling where product module
depended on order module's DTOs.
