--- src/main/java/com/ibm/demo/util/OrderItemRequest.java
@@ -1,10 +1,0 @@
-package com.ibm.demo.util;
-
-import lombok.Builder;
-
-@Builder
-public record OrderItemRequest(
-        Integer productId,
-        Integer quantity) {
-
-}
