--- src/main/java/com/ibm/demo/util/ProcessOrderItemsRequest.java
@@ -1,13 +1,0 @@
-package com.ibm.demo.util;
-
-import java.util.Set;
-
-import lombok.Builder;
-
-@Builder
-public record ProcessOrderItemsRequest(
-    Set<OrderItemRequest> originalItems,
-    Set<OrderItemRequest> updatedItems
-) {
-    
-}
